First full-cycle implementation of Datastore API.

+++ service/src/main/java/edu/lmu/cs/diabolical/ws/dao/CharacterDao.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface CharacterDao {
     List<Character> getCharacters(String name, String className, Gender gender, Integer minLevel, Integer maxLevel);
 
-    Character getCharacterById(Integer id);
+    Character getCharacterById(Long id);
 
     void deleteCharacter(Character c);
